Implement tests
This implements a (first) test to test the API, the json used comes from
a live zammad instance and has been anonymized.

It adds a Doer interface to we can mock the Do() in Send in a somewhat
elegant way.

Could make the tests fully table driven as we have some functions that
return a different number of results.

Signed-off-by: Miek Gieben <miek@miek.nl>